% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gargap-package.R
\docType{package}
\name{gargap-package}
\alias{gargap}
\alias{gargap-package}
\title{gargap: Demo Gargoyle Shiny App}
\description{
This shiny app-package was developed to demonstrate the gargoyle packge for 'Event-Based Structures for 'Shiny'.
}
\keyword{internal}
