% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gargoyleServer.R
\name{gargoyleServer}
\alias{gargoyleServer}
\title{Gargoyle Server}
\usage{
gargoyleServer(input, output, session)
}
\arguments{
\item{input}{standard shiny args}

\item{output}{standard shiny args}

\item{session}{standard shiny args}
}
\value{
shiny server.
}
\description{
Gargoyle Server
}
