% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grgylap-package.R
\docType{package}
\name{grgylap-package}
\alias{grgylap}
\alias{grgylap-package}
\title{grgylap: Demo Gargoyle Shiny App}
\description{
This shiny app-package was developed to demonstrate the gargoyle packge for 'Event-Based Structures for 'Shiny'.
}
\keyword{internal}
