% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_button.R
\name{mod_button_ui}
\alias{mod_button_ui}
\title{Gargoyle UI button function}
\usage{
mod_button_ui(id)
}
\arguments{
\item{id}{module id}
}
\value{
module ui function
}
\description{
Gargoyle UI button function
}
