% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_table.R
\name{mod_table_server}
\alias{mod_table_server}
\title{Gargoyle UI table function}
\usage{
mod_table_server(id, env_vars)
}
\arguments{
\item{id}{module id}

\item{env_vars}{output from \code{mod_button}}
}
\value{
module server function
}
\description{
Gargoyle UI table function
}
