% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_table.R
\name{mod_table_ui}
\alias{mod_table_ui}
\title{Gargoyle UI table function}
\usage{
mod_table_ui(id)
}
\arguments{
\item{id}{module id}
}
\value{
module ui function
}
\description{
Gargoyle UI table function
}
